backend/tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
coverage==7.3.2
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.2
scikit-learn==1.3.2
from app import create_app 